Make two different options to top page (IMPORTANT)
On chicken-nachos page and General Tso's page, I found two different methods to scroll back to the top.  It seems that the "#" tag almost always refers to the current page, and can be used to scroll to certain parts of the page.  "#" by itself will always scroll to the top.  This may refresh the window(?)